Commit stray diff

+++ src/net/psj/Walls/WallErase.java
@@ -5,6 +5,6 @@ import net.psj.Interface.MenuData;
 public class WallErase extends Wall {
 
 	public WallErase() {
-		super("ERSE", new float[] { 0.0f, 0.0f, 0.0f }, 0, MenuData.MENU_WALLS);
+		super("ERSE", new float[] { 0.0f, 0.0f, 0.0f }, (short) 0, MenuData.MENU_WALLS);
 	}
 }
